The long roads measure their progress without losing the horizon (server, GDD §9, travel boundary)
Regional travel calculated progress by multiplying elapsed u64 ticks by 100, which could overflow on a very long or malformed route before the bounded percentage was produced. The percentage now uses a u128 intermediate, preserving the 0-to-100 result across the full persisted tick range. A focused travel fixture verifies a half-complete maximum-length route remains at 49 percent without overflow.

+++ server/src/repository/phase5/logic.rs
@@ -259,7 +259,7 @@ pub(super) fn advance_travel(state: &mut RepositoryState) {
         }
         let total = travel.eta_tick.saturating_sub(travel.departure_tick).max(1);
         let elapsed = state.tick.saturating_sub(travel.departure_tick).min(total);
-        travel.progress = ((elapsed * 100) / total).min(100) as u8;
+        travel.progress = ((u128::from(elapsed) * 100) / u128::from(total)).min(100) as u8;
         if state.tick >= travel.eta_tick {
             travel.status = TravelStatus::Arrived;
             travel.progress = 100;
